More practice
I've learned how dictionaries work and the methods used.

I've learned 2D collections before learning the dictionary so it made it rather easier to understand.

I've created a quiz game that uses a Tuple to display the questions and 2D collection to display the options that each question has, I've also using a for loop that gets every question in the questions collection and prints them as well as a nested for loop that gets the option in options with the question number which was a variable already set in place. then I print the options, you'd have to look at the code to understand what I'm talking about but I had so much fun working on it.

I've also created a concession stand program that uses a dictionary to store items and prices. I've declared cart which contains an empty list and a total set to zero. I then use a for look to get the key, value in menu.items using the .items method in dictionaries. then I printed using an f-string  key:10  and value with a decimal of 2 spaces.

I can't explain it all here but if you take a look you can see it definitely kept me interested in what I was learning whether that be dictionaries, 2d collections or even using for or while loops which I also used in the concession stand program.